+++ dll/win32/msi/action.c
@@ -2581,7 +2581,7 @@ static BYTE *parse_value( MSIPACKAGE *package, const WCHAR *value, DWORD len, DW
                 p++;
             }
             if (deformated[0] == '-')
-                d = -d;
+                d = -(LONG)d;
             *(LPDWORD)data = d;
             TRACE("DWORD %i\n",*(LPDWORD)data);
